ADD limites e taxas nas funcoes de debito e transferencia

+++ biblioteca.h
@@ -9,7 +9,7 @@ struct contas{
     char nome[100];
     long cpf;
     int tipo_conta;
-    int valor_inicial;
+    double valor_inicial;
     char senha[100];
 };
